qa_synthetic/requirements.txt
fastapi
uvicorn
pymongo
openai
python-dotenv